Reenable signing and try to fix issues
We don't repackage bc bundles. The repackaging seems to be the
cause of classloading issues. The original provided osgi manifest works correctly.

However, version 1.57 contains an invalid BREE in its manifest which makes
tycho not happy (see https://github.com/bcgit/bc-java/issues/173).
Luckily 1.59 doesn't have that problem anymore. So we adjusted the
pom so it uses 1.59 instead and make sure we skip repackaging and signing
on 1.59.

We also have to disable automatic treatment of transitive dependencies
for sshj to avoid still pulling 1.57. And so we also have to manually include
the other dependencies for sshj.